+++ main.go
@@ -1,14 +1,16 @@
 package main
 
 import (
-	"fmt"
+	"net/http"
 	"os"
 
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
-	"github.com/iver-wharf/wharf-api-client-go/pkg/wharfapi"
+	"github.com/iver-wharf/wharf-core/pkg/ginutil"
+	"github.com/iver-wharf/wharf-core/pkg/logger"
+	"github.com/iver-wharf/wharf-core/pkg/logger/consolepretty"
 	"github.com/iver-wharf/wharf-provider-gitlab/docs"
-	log "github.com/sirupsen/logrus"
+	"github.com/iver-wharf/wharf-provider-gitlab/internal/httputils"
 	swaggerFiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
 )
@@ -20,6 +22,15 @@ const BuildDefinitionFileName = ".wharf-ci.yml"
 // ProviderName is a provider name that is used in whole wharf system for GitLab.
 const ProviderName = "gitlab"
 
+const (
+	exitCodeFailLoadVersionFile = 1
+	exitCodeFailLoadConfigFile  = 1
+	exitCodeFailLoadCerts       = 1
+	exitCodeFailBindAddress     = 2
+)
+
+var log = logger.NewScoped("WHARF-PROVIDER-GITLAB")
+
 // @title Wharf provider API for GitLab
 // @description Wharf backend API for integrating GitLab repositories with
 // @description the Wharf main API.
@@ -30,64 +41,58 @@ const ProviderName = "gitlab"
 // @contact.email wharf@iver.se
 // @basePath /import
 func main() {
-	if err := loadEmbeddedVersionFile(); err != nil {
-		fmt.Println("Failed to read embedded version.yaml file:", err)
-		os.Exit(1)
+	logger.AddOutput(logger.LevelDebug, consolepretty.Default)
+
+	var (
+		config Config
+		err    error
+	)
+	if err = loadEmbeddedVersionFile(); err != nil {
+		log.Error().WithError(err).Message("Failed to read embedded version.yaml.")
+		os.Exit(exitCodeFailLoadVersionFile)
+	}
+	if config, err = loadConfig(); err != nil {
+		log.Error().WithError(err).Message("Failed to read config.")
+		os.Exit(exitCodeFailLoadConfigFile)
 	}
 
 	docs.SwaggerInfo.Version = AppVersion.Version
 
-	initLogger(log.TraceLevel)
+	if config.CA.CertsFile != "" {
+		client, err := httputils.NewClientWithCerts(config.CA.CertsFile)
+		if err != nil {
+			log.Error().WithError(err).Message("Failed to get net/http.Client with certs.")
+			os.Exit(exitCodeFailLoadCerts)
+		}
+		http.DefaultClient = client
+	}
 
-	r := gin.Default()
+	gin.DefaultWriter = ginutil.DefaultLoggerWriter
+	gin.DefaultErrorWriter = ginutil.DefaultLoggerWriter
 
-	allowCors, ok := os.LookupEnv("ALLOW_CORS")
-	if ok && allowCors == "YES" {
-		log.Infof("Allowing CORS\n")
+	r := gin.New()
+	r.Use(
+		ginutil.DefaultLoggerHandler,
+		ginutil.RecoverProblem,
+	)
+
+	if config.HTTP.CORS.AllowAllOrigins {
+		log.Info().Message("Allowing all origins in CORS.")
 		r.Use(cors.Default())
 	}
 
 	r.GET("/", func(c *gin.Context) { c.JSON(200, gin.H{"message": "pong"}) })
-	r.POST("/import/gitlab", runGitLabHandler)
 	r.POST("/import/gitlab/trigger", runGitLabTriggerHandler)
 	r.GET("/import/gitlab/version", getVersionHandler)
 	r.GET("/import/gitlab/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
 
-	err := r.Run(getBindAddress())
-	if err != nil {
-		log.Infof("unable to run gin, error: %+v\n", err)
-	}
-}
+	importModule{&config}.register(r)
 
-func getBindAddress() string {
-	bindAddress, isBindAddressDefined := os.LookupEnv("BIND_ADDRESS")
-	if !isBindAddressDefined || bindAddress == "" {
-		return "0.0.0.0:8080"
-	}
-	return bindAddress
-}
-
-func initLogger(level log.Level) {
-	log.SetFormatter(&log.TextFormatter{
-		ForceColors:               true,
-		DisableColors:             false,
-		EnvironmentOverrideColors: false,
-		DisableTimestamp:          false,
-		FullTimestamp:             true,
-		TimestampFormat:           "",
-		DisableSorting:            false,
-		SortingFunc:               nil,
-		DisableLevelTruncation:    false,
-		QuoteEmptyFields:          false,
-		FieldMap:                  nil,
-		CallerPrettyfier:          nil,
-	})
-	log.SetLevel(level)
-}
-
-func newWharfClient(authHeader string) wharfapi.Client {
-	return wharfapi.Client{
-		ApiUrl:     os.Getenv("WHARF_API_URL"),
-		AuthHeader: authHeader,
+	if err := r.Run(config.HTTP.BindAddress); err != nil {
+		log.Error().
+			WithError(err).
+			WithString("address", config.HTTP.BindAddress).
+			Message("Failed to start web server.")
+		os.Exit(exitCodeFailBindAddress)
 	}
 }